refactor: move schema reading logic into package
