+++ packages/@cdktf/commons/src/index.ts
@@ -1,18 +1,17 @@
 // Copyright (c) HashiCorp, Inc
 // SPDX-License-Identifier: MPL-2.0
-// This is the programmatic entrypoint that the CLI uses.
-// While this is the closest we have to a programmatic API, please understand that the interfaces in this file are not stable.
-// Convert is not included here since it's published independently as @cdktf/hcl2cdk.
-/* eslint-disable @typescript-eslint/no-non-null-assertion */
 
 export * from "./checkpoint";
+export * from "./construct-maker-target";
 export * from "./config";
 export * from "./debug";
 export * from "./environment";
 export * from "./errors";
+export * from "./gradle";
 export * from "./logging";
+export * from "./module-schema";
+export * from "./provider-schema";
 export * from "./terraform-module";
 export * from "./terraform";
 export * from "./util";
 export * from "./version";
-export * from "./gradle";
