+++ packages/@cdktf/provider-generator/lib/get/__tests__/read-schema.test.ts
@@ -6,12 +6,10 @@ import {
   readModuleSchema,
 } from "../generator/provider-schema";
 import {
-  ConstructsMakerModuleTarget,
-  ConstructsMakerProviderTarget,
-} from "../constructs-maker";
-import {
   TerraformModuleConstraint,
   TerraformProviderConstraint,
+  ConstructsMakerModuleTarget,
+  ConstructsMakerProviderTarget,
   Language,
 } from "@cdktf/commons";
