+++ packages/@cdktf/provider-schema/.eslintignore
@@ -1,0 +1,2 @@
+# ignore build output
+*.d.ts
