+++ packages/@cdktf/provider-schema/.gitignore
@@ -1,0 +1,5 @@
+!tsconfig.json
+.yalc
+yalc.lock
+bundle
+!ambient.d.ts
