+++ packages/@cdktf/provider-schema/.npmignore
@@ -1,0 +1,17 @@
+*.ts
+
+!*.d.ts
+!*.js
+
+test
+dist
+.eslintignore
+.gitignore
+package.sh
+tsconfig.json
+tsconfig.tsbuildinfo
+dist
+coverage
+
+# include all files from templates
+!templates/**/*
