+++ packages/@cdktf/provider-schema/README.md
@@ -1,0 +1,3 @@
+# @cdktf/commons
+
+This package is meant to contain all code shared across the packages, like logging, types, configuration, etc. It is not meant for public usage and will not have a stable API.
