+++ packages/@cdktf/provider-schema/jest.config.js
@@ -1,0 +1,19 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+module.exports = {
+  roots: [
+    "<rootDir>"
+  ],
+  testMatch: ['**/*.test.ts', '**/*.test.tsx'],
+  transform: {
+    "^.+\\.tsx?$": "ts-jest"
+  },
+  moduleFileExtensions: [
+    "js",
+    "ts",
+    "tsx"
+  ],
+}
