+++ packages/@cdktf/provider-schema/src/index.ts
@@ -1,0 +1,1 @@
+export { readSchema } from "./read";
